1.0.6.6-SNAPSHOT

+++ src/main/java/com/sitrica/japson/shared/ReturnablePacket.java
@@ -10,7 +10,7 @@ public abstract class ReturnablePacket<T> extends Packet {
 
 	private JsonElement returned;
 
-	public ReturnablePacket(byte id) {
+	public ReturnablePacket(int id) {
 		super(id);
 	}
